split algae gripper subsystem
