+++ src/main/java/frc/robot/auto/BlueL4AlgaeAuto.java
@@ -10,7 +10,7 @@ import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.Constants.ElevatorConstants.ElevatorState;
 import frc.robot.drivetrain.Drivetrain;
 import frc.robot.elevator.Elevator;
-import frc.robot.grippers.AlgaeIntake;
+import frc.robot.grippers.AlgaeRoller;
 import frc.robot.grippers.CoralRoller;
 import java.util.Optional;
 
@@ -18,14 +18,14 @@ public class BlueL4AlgaeAuto extends AutoMode {
 
   Elevator elevator;
   CoralRoller coralRoller;
-  AlgaeIntake algaeIntake;
+  AlgaeRoller algaeIntake;
   AutoCGs autoCG;
 
   public BlueL4AlgaeAuto(
       Drivetrain drivetrain,
       Elevator elevatorsubsystem,
       CoralRoller coralRollerSubsystem,
-      AlgaeIntake algaeIntakeSubsystem,
+      AlgaeRoller algaeIntakeSubsystem,
       AutoCGs autoCommandGroups) {
     super(drivetrain);
     elevator = elevatorsubsystem;
